requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.11
mysql_connector_repackaged==0.3.1
pandas==2.2.3
pydantic==2.10.6
PyJWT==2.10.1
pymongo==4.11.2
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
